Remove schema property in favor of get_schema method
Previous implementation relied on a client being instantiated within
an asyncio context so as to enable automated introspection. This is no
longer required, and should be done lazily on the first request.

In order to make this behavior change, access to the `schema` property
has been removed in favour or `get_shema`.